Minor changes have been made.
