+++ src/app/pages/roles-list/create-role-dialog/create-role-dialog.component.ts
@@ -18,7 +18,6 @@ import {
 import {
   FormControl,
   FormsModule,
-  Validators,
   ReactiveFormsModule,
 } from '@angular/forms';
